calculo dos limites e apresentação em grafico

+++ ProjetoCEP/Projeto.Entidades/Lote.cs
@@ -15,7 +15,7 @@ namespace Projeto.Entidades
         public Operador OperadorAnalise { get; set; }
         public int TotalLentes { get; set; }
         public int QtdNaoConforme { get; set; }
-        public decimal Percentual { get; set; }
+        public double Percentual { get; set; }
         public Status Status { get; set; }
         public LimitesControle Limites { get; set; }
         public TipoLote TipoLote { get; set; }
